Live Feed Fixed

+++ jetPackPam/app/src/main/java/student/projects/jetpackpam/data/CameraRequest.kt
@@ -1,6 +1,6 @@
 package student.projects.jetpackpam.data
 
 data class CameraRequest(
-    val pan: Int,
-    val tilt: Int
+    val pan: Float,
+    val tilt: Float
 )
